refactored addOrder. Api call is working
